|Amey|Mod|notification chnages review|

+++ src/main/java/com/lti/mod/services/notificationservice/repository/NotificationRepository.java
@@ -10,5 +10,5 @@ import org.springframework.data.repository.CrudRepository;
 public interface NotificationRepository extends CrudRepository<NotificationDO,Integer> {
 
     @Query("SELECT u FROM NotificationDO u WHERE u.userId = ?1")
-    public List<NotificationDO> getNotificationByUserId(Integer userId);
+    public List<NotificationDO> getNotificationByUserId(Long userId);
 }
